Add files via upload
The segment_image() function takes an image I as an input which is already read in the “double” format 
from integer format, meaning it is scaled and ready for all the transformations. 
Next the image is transformed into a grayscale format, making the third dimension equal to 1 instead of 3 
(which is the case for colourful images in the RGB format). Next image gets convolved with a Gaussian mask 
with a standard deviation of 2 and a mask size of 12 for initial smoothing using the function “conv(img)”. 
Afterward the Image undergoes Fuzzy Logic Edge Detection, which is an effective edge detection technique 
that uses membership functions to define specific degrees regarding whether a pixel belongs to an edge or a uniform region. 
This method uses the image gradient to detect local breaks. Next, the original image is convolved 
with the defined gradients and together with a set of other parameters (such as sx or sy, 
which resemble the standard deviation for the zero membership function), all the relevant matrices 
and parameters are added to the Fuzzy Inference System (FIS). Next, the Fuzzy Logic Edge Detection method 
returns an image with highly detailed edges. Because the returned images have so much detail, the output
 of the aforementioned method goes through another edge detection function "edges(img)" - which uses the
 Canny Edge detection method on our image with very detailed images, making them slightly less detailed
 and making them more look like the edges where drawn with "A thick line”. The reason for this
 additional edge detection is that the training sets images with edges drawn by humans were reasonably thick, 
therefore in order to achieve images resembling the ones that humans have drawn, the last step of the 
segment_image(I) function was essential.

Source : https://uk.mathworks.com/help/fuzzy/fuzzy-logic-image-processing.html